fix:review structure app property

+++ src/main/kotlin/server/web/casa/app/property/infrastructure/persistence/PropertyFavoriteEntity.kt
@@ -8,11 +8,7 @@ data class PropertyFavoriteEntity(
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val propertyFavoriteId : Int,
-//    @Column(name = "userId")
-//    val userId : Int,
-//    @Column(name = "propertyId")
-//    val propertyId : Int,
+    val propertyFavoriteId : Long,
     @ManyToMany(mappedBy = "favorites")
     val property : List<PropertyEntity>
 )
